Subset of the data (cleaned data)
First, we will change the temperate from Celsius to Fahrenheit as we are presenting to an
American audience. We do not believe that changing the meters units is necessary as these are
more easily relatable. After that we will make dummy indicator columns for our categorical
variables of season, holiday, and functioning day.
After that we want to make a new column called “Good hour” this column will be a categorical
column that will tell us if the hour has a good temperature (40-85), humidity (30-60), wind speed
(less than or equal to 8), visibility (greater than or equal to 1000), rainfall (0), and snowfall (0).
Consult appendix 3 for good hour visualization.
We will also convert the date variable into a categorical to describe if the day is a weekend or
not. If we converted this into each day of the week that would create too much noise for our
model, and we believe that there would only be a significant difference between the week and
weekend. Now the date variable is being accurately described by our weekend and season
variables, and we will not include date it in the model.
Before we model, we do not expect there to be any target leakage as there are no variables that
would be perfect predictors of the rented bike count. We also expect some multicollinearity
between variables that are related to weather as those variables can be highly correlated.